Create src directory
