+++ index.js
@@ -4,8 +4,9 @@ const discoverySwarm = require('discovery-swarm')
 const swarmDefaults = require('dat-swarm-defaults')
 const Debug = require('debug')
 
-const Multicore = require('./multicore')
-const Hyperfile = require('./hyperfile')
+const Multicore = require('./src/multicore')
+const Hyperfile = require('./src/hyperfile')
+const DocHandle = require('./src/doc-handle')
 
 const log = Debug('hypermerge:index')
 
@@ -30,72 +31,6 @@ const METADATA = {
  * An Automerge document.
  * @typedef {object} Document
  */
-
-/**
-   * Experimental second open
-   */
-class DocHandle {
-  constructor (hm, docId) {
-    this.hm = hm
-    this.id = docId
-    this._cb = () => {}
-  }
-
-  get () {
-    if (this.hm.readyIndex[this.id]) {
-      return this.hm.find(this.id)
-    }
-    return null
-  }
-
-  /* make a change to a document through the handle */
-  change (cb) {
-    const doc = this.hm.find(this.id)
-    this.hm.change(doc, cb)
-    return this.hm.find(this.id)
-  }
-
-  /* register the function you'd like called when the document is updated */
-  onChange (cb) {
-    this._cb = cb
-    if (this.hm.readyIndex[this.id]) {
-      const doc = this.hm.find(this.id)
-      cb(doc)
-    }
-    return this
-  }
-
-  message (message) {
-    if (this.hm.readyIndex[this.id]) {
-      this.hm.message(this.id, message)
-    }
-  }
-
-  /* register the function you'd like called when you receive a message from a peer */
-  onMessage (cb) {
-    this._messageCb = cb
-    return this
-  }
-
-  release () {
-    this.hm.releaseHandle(this)
-  }
-
-  _message ({ peer, msg }) {
-    if (this._messageCb) {
-      this._messageCb({ peer, msg })
-    }
-  }
-
-  _update (doc) {
-    this._cb(doc)
-  }
-
-  _ready (doc) {
-    this._cb(doc)
-  }
-}
-
 /**
  * Creates a new Hypermerge instance that manages a set of documents.
  * All previously opened documents are automatically re-opened.
