+++ src/multicore.js
